Restructuring a little bit
I know it is going to exponentially add time to the program,
however, I am mainly focused on the highest bang for the buck
and if it takes an hour to compute, then it still will be worth it.

Added a list of children instead of just 4 children.  This way
I can test out every possible combination.