Still having Overlap Issues
I am starting back at the beginning again since my original
idea seems to be prone to bugs.  I still can not get the
overlapping issues on test case 12 to stop.  It is difficult
to see the exact problem when the datasets are so massive..

I'm going to start back at the drawing board.